moved load_data from monitoring.tools to src.data
